Consolidate Game Center v2 client/types.
Fold v2 achievement/leaderboard types and client methods into existing files and drop redundant v2 files.
