--- internal/asc/client_game_center_v2.go
@@ -1,219 +1,0 @@
-package asc
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"strings"
-)
-
-// GetGameCenterAchievementV2 retrieves a v2 Game Center achievement by ID.
-func (c *Client) GetGameCenterAchievementV2(ctx context.Context, achievementID string) (*GameCenterAchievementResponse, error) {
-	path := fmt.Sprintf("/v2/gameCenterAchievements/%s", strings.TrimSpace(achievementID))
-	data, err := c.do(ctx, http.MethodGet, path, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterAchievementResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// CreateGameCenterAchievementV2 creates a new v2 Game Center achievement.
-func (c *Client) CreateGameCenterAchievementV2(ctx context.Context, gcDetailID string, groupID string, attrs GameCenterAchievementCreateAttributes) (*GameCenterAchievementResponse, error) {
-	relationships := &GameCenterAchievementV2Relationships{}
-	hasRelationship := false
-
-	if strings.TrimSpace(gcDetailID) != "" {
-		relationships.GameCenterDetail = &Relationship{
-			Data: ResourceData{
-				Type: ResourceTypeGameCenterDetails,
-				ID:   strings.TrimSpace(gcDetailID),
-			},
-		}
-		hasRelationship = true
-	}
-	if strings.TrimSpace(groupID) != "" {
-		relationships.GameCenterGroup = &Relationship{
-			Data: ResourceData{
-				Type: ResourceTypeGameCenterGroups,
-				ID:   strings.TrimSpace(groupID),
-			},
-		}
-		hasRelationship = true
-	}
-	if !hasRelationship {
-		relationships = nil
-	}
-
-	payload := GameCenterAchievementV2CreateRequest{
-		Data: GameCenterAchievementV2CreateData{
-			Type:          ResourceTypeGameCenterAchievements,
-			Attributes:    attrs,
-			Relationships: relationships,
-		},
-	}
-
-	body, err := BuildRequestBody(payload)
-	if err != nil {
-		return nil, err
-	}
-
-	data, err := c.do(ctx, http.MethodPost, "/v2/gameCenterAchievements", body)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterAchievementResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// UpdateGameCenterAchievementV2 updates an existing v2 Game Center achievement.
-func (c *Client) UpdateGameCenterAchievementV2(ctx context.Context, achievementID string, attrs GameCenterAchievementUpdateAttributes) (*GameCenterAchievementResponse, error) {
-	payload := GameCenterAchievementUpdateRequest{
-		Data: GameCenterAchievementUpdateData{
-			Type:       ResourceTypeGameCenterAchievements,
-			ID:         strings.TrimSpace(achievementID),
-			Attributes: &attrs,
-		},
-	}
-
-	body, err := BuildRequestBody(payload)
-	if err != nil {
-		return nil, err
-	}
-
-	path := fmt.Sprintf("/v2/gameCenterAchievements/%s", strings.TrimSpace(achievementID))
-	data, err := c.do(ctx, http.MethodPatch, path, body)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterAchievementResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// DeleteGameCenterAchievementV2 deletes a v2 Game Center achievement.
-func (c *Client) DeleteGameCenterAchievementV2(ctx context.Context, achievementID string) error {
-	path := fmt.Sprintf("/v2/gameCenterAchievements/%s", strings.TrimSpace(achievementID))
-	_, err := c.do(ctx, http.MethodDelete, path, nil)
-	return err
-}
-
-// GetGameCenterLeaderboardV2 retrieves a v2 Game Center leaderboard by ID.
-func (c *Client) GetGameCenterLeaderboardV2(ctx context.Context, leaderboardID string) (*GameCenterLeaderboardResponse, error) {
-	path := fmt.Sprintf("/v2/gameCenterLeaderboards/%s", strings.TrimSpace(leaderboardID))
-	data, err := c.do(ctx, http.MethodGet, path, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterLeaderboardResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// CreateGameCenterLeaderboardV2 creates a new v2 Game Center leaderboard.
-func (c *Client) CreateGameCenterLeaderboardV2(ctx context.Context, gcDetailID string, groupID string, attrs GameCenterLeaderboardCreateAttributes) (*GameCenterLeaderboardResponse, error) {
-	relationships := &GameCenterLeaderboardV2Relationships{}
-	hasRelationship := false
-
-	if strings.TrimSpace(gcDetailID) != "" {
-		relationships.GameCenterDetail = &Relationship{
-			Data: ResourceData{
-				Type: ResourceTypeGameCenterDetails,
-				ID:   strings.TrimSpace(gcDetailID),
-			},
-		}
-		hasRelationship = true
-	}
-	if strings.TrimSpace(groupID) != "" {
-		relationships.GameCenterGroup = &Relationship{
-			Data: ResourceData{
-				Type: ResourceTypeGameCenterGroups,
-				ID:   strings.TrimSpace(groupID),
-			},
-		}
-		hasRelationship = true
-	}
-	if !hasRelationship {
-		relationships = nil
-	}
-
-	payload := GameCenterLeaderboardV2CreateRequest{
-		Data: GameCenterLeaderboardV2CreateData{
-			Type:          ResourceTypeGameCenterLeaderboards,
-			Attributes:    attrs,
-			Relationships: relationships,
-		},
-	}
-
-	body, err := BuildRequestBody(payload)
-	if err != nil {
-		return nil, err
-	}
-
-	data, err := c.do(ctx, http.MethodPost, "/v2/gameCenterLeaderboards", body)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterLeaderboardResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// UpdateGameCenterLeaderboardV2 updates an existing v2 Game Center leaderboard.
-func (c *Client) UpdateGameCenterLeaderboardV2(ctx context.Context, leaderboardID string, attrs GameCenterLeaderboardUpdateAttributes) (*GameCenterLeaderboardResponse, error) {
-	payload := GameCenterLeaderboardUpdateRequest{
-		Data: GameCenterLeaderboardUpdateData{
-			Type:       ResourceTypeGameCenterLeaderboards,
-			ID:         strings.TrimSpace(leaderboardID),
-			Attributes: &attrs,
-		},
-	}
-
-	body, err := BuildRequestBody(payload)
-	if err != nil {
-		return nil, err
-	}
-
-	path := fmt.Sprintf("/v2/gameCenterLeaderboards/%s", strings.TrimSpace(leaderboardID))
-	data, err := c.do(ctx, http.MethodPatch, path, body)
-	if err != nil {
-		return nil, err
-	}
-
-	var response GameCenterLeaderboardResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// DeleteGameCenterLeaderboardV2 deletes a v2 Game Center leaderboard.
-func (c *Client) DeleteGameCenterLeaderboardV2(ctx context.Context, leaderboardID string) error {
-	path := fmt.Sprintf("/v2/gameCenterLeaderboards/%s", strings.TrimSpace(leaderboardID))
-	_, err := c.do(ctx, http.MethodDelete, path, nil)
-	return err
-}
