+++ internal/asc/game_center_leaderboards.go
@@ -59,6 +59,12 @@ type GameCenterLeaderboardRelationships struct {
 	GameCenterDetail *Relationship `json:"gameCenterDetail"`
 }
 
+// GameCenterLeaderboardV2Relationships describes relationships for v2 leaderboards.
+type GameCenterLeaderboardV2Relationships struct {
+	GameCenterDetail *Relationship `json:"gameCenterDetail,omitempty"`
+	GameCenterGroup  *Relationship `json:"gameCenterGroup,omitempty"`
+}
+
 // GameCenterLeaderboardCreateData is the data portion of a leaderboard create request.
 type GameCenterLeaderboardCreateData struct {
 	Type          ResourceType                          `json:"type"`
@@ -69,6 +75,18 @@ type GameCenterLeaderboardCreateData struct {
 // GameCenterLeaderboardCreateRequest is a request to create a leaderboard.
 type GameCenterLeaderboardCreateRequest struct {
 	Data GameCenterLeaderboardCreateData `json:"data"`
+}
+
+// GameCenterLeaderboardV2CreateData is the data portion of a v2 leaderboard create request.
+type GameCenterLeaderboardV2CreateData struct {
+	Type          ResourceType                          `json:"type"`
+	Attributes    GameCenterLeaderboardCreateAttributes `json:"attributes"`
+	Relationships *GameCenterLeaderboardV2Relationships `json:"relationships,omitempty"`
+}
+
+// GameCenterLeaderboardV2CreateRequest is a request to create a v2 leaderboard.
+type GameCenterLeaderboardV2CreateRequest struct {
+	Data GameCenterLeaderboardV2CreateData `json:"data"`
 }
 
 // GameCenterLeaderboardUpdateData is the data portion of a leaderboard update request.
